chore(backend): make the Sync client non-nullable, collapse the absent-sync seam
Sync is required config — the process exits at startup without
SYNC_SERVICE_URL/SYNC_INTERNAL_AUTH_TOKEN — yet getSyncServiceClient()
still returned SyncServiceClient | null and nine call sites re-checked
for a null that cannot happen in a running process, including a 22-line
fabricated-metadata fallback in user-metadata.service.ts with its own
dedicated test file.

- sync-service.factory.ts: getSyncServiceClient() now returns a real
  client unconditionally.
- Removes the null guards in event.controller.ts, calendar.controller.ts
  (x2), auth.controller.ts, user.service.ts#purgeSyncPrincipal, and
  sync-change-feed.bridge.ts (x2), plus the now-dead
  AuthError.ConnectNotDelegated (and two pre-existing dead entries,
  MissingRefreshToken/NoUserId, noticed in the same file).
- Deletes assessGoogleMetadata's local fallback entirely along with its
  isolated test file.
- Deletes the sync-service.client.ts methods with zero backend callers:
  listEventOccurrences (superseded by listFullEvents) and
  resolveDiagnosticConnection (the sync-side diagnostic route is meant
  to be curled directly by operators, not through this wrapper) — and,
  cross-package, the now-unreachable GET /internal/events route +
  toSyncEventOccurrence in packages/sync, plus the EventOccurrenceList*/
  SyncEventOccurrence schemas in core.
- Deletes the deprecated resolveGoogleConnectionStateFromSync wrapper,
  folding its still-valuable test coverage into
  resolveGoogleConnectionFromSync's tests.
- Rewrites the two *.delegation.test.ts files, which tested "the sync
  branch ran" against a fork that no longer exists, as plain
  fails-closed-on-outage tests; merges calendar's into the existing
  calendar.controller.test.ts, renames event's to event.controller.test.ts
  (no prior plain test file existed for it).
- Fixes compass.example.yaml, which commented out the entire sync: block
  as "optional" and would leave a self-hoster's backend refusing to
  boot; also drops the channelExpirationMin/webhookUrl/notificationToken
  lines under google: that PR2 already removed from the schema.

Left alone on purpose: sync-connection-begin.ts has real translation
logic (error-kind mapping, correlation-id logging) with clean isolated
unit tests — inlining it into auth.controller.ts would lose that test
isolation for a marginal LOC win, so it stays as its own module.

Part 3 of the post-Sync-cutover simplification pass (see #2484, #2486).
Zero behavior change: backend/core/sync suites match their pre-change
pass counts exactly (dropped tests are 1:1 accounted for by the
intentional deletions/consolidations above; sync and core show 0 fail).
