+++ packages/backend/src/common/errors/auth/auth.errors.ts
@@ -2,26 +2,17 @@ import { Status } from "@core/errors/status.codes";
 import { type ErrorMetadata } from "@backend/common/types/error.types";
 
 interface AuthErrors {
-  ConnectNotDelegated: ErrorMetadata;
   DevOnly: ErrorMetadata;
   GoogleAccountAlreadyConnected: ErrorMetadata;
   GoogleConnectEmailMismatch: ErrorMetadata;
   GoogleNotConfigured: ErrorMetadata;
   GoogleRedirectUriMismatch: ErrorMetadata;
   InadequatePermissions: ErrorMetadata;
-  MissingRefreshToken: ErrorMetadata;
-  NoUserId: ErrorMetadata;
   NoGAuthAccessToken: ErrorMetadata;
   SyncConnectionUnavailable: ErrorMetadata;
 }
 
 export const AuthError: AuthErrors = {
-  ConnectNotDelegated: {
-    description:
-      "Provider connect is not delegated to the sync service on this deployment",
-    status: Status.CONFLICT,
-    isOperational: true,
-  },
   DevOnly: {
     description: "Only available during development",
     status: Status.FORBIDDEN,
@@ -55,16 +46,6 @@ export const AuthError: AuthErrors = {
     description: "You don't have permission to do that",
     status: Status.FORBIDDEN,
     isOperational: true,
-  },
-  MissingRefreshToken: {
-    description: "No refresh token",
-    status: Status.BAD_REQUEST,
-    isOperational: true,
-  },
-  NoUserId: {
-    description: "Compass user was not created",
-    status: Status.INTERNAL_SERVER,
-    isOperational: false,
   },
   NoGAuthAccessToken: {
     description: "No gauth access token",
