+++ packages/sync/src/safety/isolation-matrix.db.test.ts
@@ -13,7 +13,6 @@ import { COMMANDS_PATH } from "@sync/server/command.routes";
 import {
   AVAILABILITY_BUSY_PATH,
   CONNECTIONS_PATH,
-  EVENTS_PATH,
 } from "@sync/server/connection.routes";
 import { PRINCIPAL_PATH } from "@sync/server/principal.routes";
 import { SYNC_COLLECTIONS } from "@sync/storage/collections";
@@ -567,32 +566,6 @@ describe("R-SEC-03 isolation matrix", () => {
         ),
       ).not.toBeNull();
       expect(mine).toBeDefined();
-    });
-
-    it("does not list another principal's events for a shared calendar id shape", async () => {
-      const tenantId = objectId();
-      const owner = objectId();
-      const attacker = objectId();
-      await startService();
-      const connectionId = await seedConnection(tenantId, owner);
-      const calendarId = await seedBusyCalendar(tenantId, owner, connectionId);
-
-      const res = await fetch(
-        `${base}${EVENTS_PATH}?calendarIds=${calendarId}&start=2026-07-13T00:00:00.000Z&end=2026-07-15T00:00:00.000Z`,
-        { headers: signedHeaders(tenantId, attacker) },
-      );
-      expect(res.status).toBe(200);
-      const body = (await res.json()) as { occurrences: unknown[] };
-      expect(body.occurrences).toEqual([]);
-
-      const ownerPage = (await (
-        await fetch(
-          `${base}${EVENTS_PATH}?calendarIds=${calendarId}&start=2026-07-13T00:00:00.000Z&end=2026-07-15T00:00:00.000Z`,
-          { headers: signedHeaders(tenantId, owner) },
-        )
-      ).json()) as { occurrences: Array<{ title: string }> };
-      expect(ownerPage.occurrences).toHaveLength(1);
-      expect(ownerPage.occurrences[0]?.title).toBe("owner-secret-meeting");
     });
   });
